+++ tests/e2e/cases/async_await_while_condition/in.ts
@@ -109,6 +109,14 @@ async function chooseDirectAwaitUninitializedVarForPrelude(prefix: string): Prom
     return await laterBodyValue(source);
 }
 
+async function chooseDirectAwaitUninitializedVarForOfPrelude(prefix: string): Promise<string> {
+    var source: any;
+    for (const item of ["of"]) {
+        source = prefix + "-" + item;
+    }
+    return await laterBodyValue(source);
+}
+
 class MutableUninitializedVarAwaitChooser {
     async choose(prefix: string): Promise<string> {
         var source: string;
@@ -3764,6 +3772,7 @@ chooseDirectAwaitUninitializedVarIfPrelude("direct-uninitialized-var-if", false)
 chooseDirectAwaitUninitializedVarWhilePrelude("direct-uninitialized-var-while", true).then((value) => console.log("await-direct-uninitialized-var-while-true", value));
 chooseDirectAwaitUninitializedVarDoPrelude("direct-uninitialized-var-do").then((value) => console.log("await-direct-uninitialized-var-do", value));
 chooseDirectAwaitUninitializedVarForPrelude("direct-uninitialized-var-for").then((value) => console.log("await-direct-uninitialized-var-for", value));
+chooseDirectAwaitUninitializedVarForOfPrelude("direct-uninitialized-var-for-of").then((value) => console.log("await-direct-uninitialized-var-for-of", value));
 new MutableUninitializedVarAwaitChooser().choose("method-mutable-uninitialized-var").then((value) => console.log("await-method-mutable-uninitialized-var-prelude", value));
 new MutableUninitializedVarAwaitChooser().chooseThrow("method-mutable-uninitialized-var").catch((reason) => console.log("await-method-mutable-uninitialized-var-throw", reason));
 chooseMutableUninitializedVarAwaitValue("value-mutable-uninitialized-var").then((value) => console.log("await-value-mutable-uninitialized-var-prelude", value));